server/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.7.0
attrs==24.2.0
autopep8==2.3.1
bcrypt==4.1.3
black==24.10.0
blinker==1.8.2
boto3==1.35.39
botocore==1.35.39
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
filelock==3.16.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.10
frozenlist==1.5.0
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.2
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.66.1
grpcio-status==1.66.1
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.1.2
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.8.0
PyMySQL==1.1.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rsa==4.9
s3transfer==0.10.3
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.6.0
six==1.16.0
SQLAlchemy==2.0.30
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.5.1
torch-geometric==2.6.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.0.3
yarl==1.18.0
